Migrate Redis Sink in databases-jvm
